edit run

+++ testing/run.sh
@@ -1,10 +1,26 @@
 #!/bin/bash
 
-curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/reads.bam
-curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/reads.bam.bai
-#
-curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/test.bam
-curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/test.bam.bai
-#
-curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/test.vcf.gz
-curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/test.vcf.gz.tbi
+
+mkdir -p input_data
+cd input_data
+
+if [ ! -s reads.bam.bai ]
+then
+    curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/reads.bam
+    curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/reads.bam.bai
+fi
+
+if [ ! -s test.bam.bai ]
+then
+    curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/test.bam
+    curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/test.bam.bai
+fi
+
+if [ ! -s test.vcf.gz.tbi ]
+then
+    curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/test.vcf.gz
+    curl -O https://rodrio10.u.hpc.mssm.edu/MsPAC/test.vcf.gz.tbi
+fi
+
+cd -
+
